trua 23.11.2021

+++ database/migrations/2021_06_15_000004_create_donhangs_table.php
@@ -34,7 +34,7 @@ class CreateDonhangsTable extends Migration
             $table->bigInteger('tongchiphi');
             $table->bigInteger('chietkhau');
             $table->text('ghichu')->nullable();     
-            $table->integer('tongkhoiluong')->nullable(); 
+            $table->float('tongkhoiluong', 8, 2)->nullable(); 
             $table->timestamps();
         });
     }
